# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Task Manager API",
    "description": "Personal task management \u2014 single-user, local-only.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/goals": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "List Goals",
        "operationId": "list_goals_goals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalResponse"
                  },
                  "type": "array",
                  "title": "Response List Goals Goals Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Create Goal",
        "operationId": "create_goal_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/goals/{goal_id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Get Goal",
        "operationId": "get_goal_goals__goal_id__get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Update Goal",
        "operationId": "update_goal_goals__goal_id__patch",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete Goal",
        "operationId": "delete_goal_goals__goal_id__delete",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/goals/{goal_id}/key-results": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "List Key Results",
        "operationId": "list_key_results_goals__goal_id__key_results_get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KeyResultResponse"
                  },
                  "title": "Response List Key Results Goals  Goal Id  Key Results Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Create Key Result",
        "operationId": "create_key_result_goals__goal_id__key_results_post",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyResultCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/goals/{goal_id}/key-results/{kr_id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Get Key Result",
        "operationId": "get_key_result_goals__goal_id__key_results__kr_id__get",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Kr Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Update Key Result",
        "operationId": "update_key_result_goals__goal_id__key_results__kr_id__patch",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Kr Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyResultUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete Key Result",
        "operationId": "delete_key_result_goals__goal_id__key_results__kr_id__delete",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          },
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Kr Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/burndown": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Burndown",
        "operationId": "get_burndown_dashboard_burndown_get",
        "parameters": [
          {
            "name": "sprint_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sprint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BurndownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/velocity": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Velocity",
        "operationId": "get_velocity_dashboard_velocity_get",
        "parameters": [
          {
            "name": "last_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 5,
              "title": "Last N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VelocityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BurndownPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "remaining_points": {
            "type": "integer",
            "title": "Remaining Points"
          }
        },
        "type": "object",
        "required": [
          "date",
          "remaining_points"
        ],
        "title": "BurndownPoint"
      },
      "BurndownResponse": {
        "properties": {
          "sprint_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sprint Id"
          },
          "sprint_name": {
            "type": "string",
            "title": "Sprint Name"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "total_points": {
            "type": "integer",
            "title": "Total Points"
          },
          "ideal_line": {
            "items": {
              "$ref": "#/components/schemas/BurndownPoint"
            },
            "type": "array",
            "title": "Ideal Line"
          },
          "actual_remaining": {
            "type": "integer",
            "title": "Actual Remaining"
          }
        },
        "type": "object",
        "required": [
          "sprint_id",
          "sprint_name",
          "start_date",
          "end_date",
          "total_points",
          "ideal_line",
          "actual_remaining"
        ],
        "title": "BurndownResponse"
      },
...
      "GoalCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Priority",
            "default": "medium"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "date_range_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "date_range_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "GoalCreateRequest"
      },
...
      "GoalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "date_range_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "date_range_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "priority",
          "tags",
          "date_range_start",
          "date_range_end",
          "created_at",
          "updated_at"
        ],
        "title": "GoalResponse"
      },
      "GoalUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "critical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "date_range_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "date_range_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          }
        },
        "type": "object",
        "title": "GoalUpdateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KeyResultCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "target_value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Target Value"
          },
          "current_value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Current Value",
            "default": 0.0
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "target_value"
        ],
        "title": "KeyResultCreateRequest"
      },
      "KeyResultResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "goal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Goal Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "target_value": {
            "type": "number",
            "title": "Target Value"
          },
          "current_value": {
            "type": "number",
            "title": "Current Value"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "progress_percent": {
            "type": "number",
            "title": "Progress Percent"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal_id",
          "title",
          "description",
          "target_value",
          "current_value",
          "unit",
          "progress_percent",
          "created_at",
          "updated_at"
        ],
        "title": "KeyResultResponse"
      },
      "KeyResultUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "current_value": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Value"
          },
          "target_value": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "type": "object",
        "title": "KeyResultUpdateRequest"
      },
...
      "SprintVelocity": {
        "properties": {
          "sprint_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sprint Id"
          },
          "sprint_name": {
            "type": "string",
            "title": "Sprint Name"
          },
          "completed_points": {
            "type": "integer",
            "title": "Completed Points"
          }
        },
        "type": "object",
        "required": [
          "sprint_id",
          "sprint_name",
          "completed_points"
        ],
        "title": "SprintVelocity"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VelocityResponse": {
        "properties": {
          "sprints": {
            "items": {
              "$ref": "#/components/schemas/SprintVelocity"
            },
            "type": "array",
            "title": "Sprints"
          },
          "average_velocity": {
            "type": "number",
            "title": "Average Velocity"
          }
        },
        "type": "object",
        "required": [
          "sprints",
          "average_velocity"
        ],
        "title": "VelocityResponse"
      }
    }
  }
}